% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{binom_prob}
\alias{binom_prob}
\title{Binomial Probability}
\usage{
binom_prob(n_success, prob_success, sample_size)
}
\arguments{
\item{n_success}{number of successes to calculate}

\item{prob_success}{probability of a single success}

\item{sample_size}{size of the sample (number of trials)}
}
\description{
Binomial Probability
}
\examples{
binom_prob(n_success = 10, prob_success = 0.5, sample_size = 20)
}
