% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cards}
\alias{cards}
\title{A Standard Set of Playing Cards}
\usage{
cards(jokers = FALSE, makespace = FALSE)
}
\arguments{
\item{jokers}{logical. Include jokers in the deck.}

\item{makespace}{logical. Include a column of equally likely probabilities.}
}
\value{
A data frame with columns \code{rank} and \code{suit}, and optionally a column of equally likely \code{probs}.
}
\description{
The title says it all.
}
\details{
This generates a data frame sample space of a standard deck of 52 playing cards. Optionally, the user can specify that Jokers be included, which have a \code{rank} but with \code{suit} a missing value.
}
\examples{
cards()
cards(makespace = TRUE)

}
\seealso{
\code{\link[probs]{rolldie}}, \code{\link[probs]{tosscoin}}, and \code{\link[probs]{roulette}}
}
\author{
G. Jay Kerns \email{gkerns@ysu.edu}.
}
\keyword{data}
