% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fAsianOptions.R
\name{cgamma}
\alias{cgamma}
\title{A function implemented by Diethelm Wuertz}
\source{
For the Fortran Routine: http://iris-lee3.ece.uiuc.edu/~jjin/routines/routines.html
}
\usage{
cgamma(x, log = FALSE)
}
\arguments{
\item{x}{a complex or real vector}

\item{log}{if TRUE the logarithm of the gamma is calculated otherwise if FALSE, the gamma function itself will be calculated.}
}
\value{
a complex value
}
\description{
Computes the Gamma Function for complex argument "x"
}
\author{
Diethelm Wuertz
}
