% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{countrep.data.frame}
\alias{countrep.data.frame}
\title{Count Repetitions
Counts the number of repetitions of \code{vals} in a given vector \code{x}.}
\usage{
\method{countrep}{data.frame}(x, ...)
}
\arguments{
\item{x}{an object in which repeats should be counted.}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
If \code{x} is a vector, then the value is an integer.  If \code{x} is a data frame then the value is a vector, with entries the corresponding value for the respective rows of \code{x}
}
\description{
Count Repetitions
Counts the number of repetitions of \code{vals} in a given vector \code{x}.
}
\details{
This is a generic function, with methods supplied for data frames and vectors. The default behavior counts the number of pairs of elements of \code{x}.  One can find the number of triples, etc., by changing the \code{nrep} argument.  If there are specific values for which one is looking for repeats, these can be specified with the \code{vals} argument.  Note that the function only checks for \emph{exactly} \code{nrep} instances, so two pairs of a specific element would be counted as 0 pairs and 1 quadruple. See the examples.
}
\examples{
x <- c(3,3,2,2,3,3,4,4)
countrep(x)  # one pair each of 2s and 4s
countrep(x, nrep = 4)
countrep(x, vals = 4) # one pair of 4s
}
\seealso{
\code{\link[probs]{isrep}}
}
\author{
G. Jay Kerns \email{gkerns@ysu.edu}.
}
