% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{cumul_binom_prob}
\alias{cumul_binom_prob}
\title{Cumulative Binomial Probability}
\usage{
cumul_binom_prob(n_success, prob_success, sample_size)
}
\arguments{
\item{n_success}{cumulative number of successes to calculate}

\item{prob_success}{probability of a single success}

\item{sample_size}{size of the sample (number of trials)}
}
\description{
Cumulative Binomial Probability
}
\examples{
cumul_binom_prob(n_success = 10, prob_success = 0.5, sample_size = 20)
}
