% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genData.R
\name{genLogRegData}
\alias{genLogRegData}
\title{Generate data for logistic regression}
\usage{
genLogRegData(xdata, beta = rep(1, ncol(xdata)), yname = "y")
}
\arguments{
\item{xdata}{the model matrix}

\item{beta}{vector of parameters to multiply the model matrix}

\item{yname}{the name for the generated y values}
}
\value{
A data frame with the model matrix and the generated y values added.
}
\description{
This function generates data ready for a logistic regression model.
}
\details{
This function generates data ready for a logistic regression model.
}
\author{
G. Jay Kerns
}
\keyword{datagen}
