% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fAsianOptions.R
\name{kummerU}
\alias{kummerU}
\title{A function implemented by Diethelm Wuertz}
\usage{
kummerU(x, a, b, ip = 0)
}
\arguments{
\item{x}{complex function argument}

\item{a}{complex index}

\item{b}{complet index}

\item{ip}{default 0}
}
\value{
a complex value
}
\description{
Calculate the Confluent Hypergeometric Function of the Second Kind for complex argument "x" and complex indexes "a" and "b"
}
\author{
Diethelm Wuertz
}
